requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch
torchvision
torchaudio
tensorflow
numpy
scikit-learn
pillow
tqdm
eel
